Update PrepaidController.php, SmsController.php, and 14 more files...

+++ database/migrations/2020_12_15_133518_create_bandwidths_table.php
@@ -16,9 +16,9 @@ class CreateBandwidthsTable extends Migration
         Schema::create('bandwidths', function (Blueprint $table) {
             $table->id();
             $table->string('name',100);
-            $table->integer('rate_down');
+            $table->float('rate_down');
             $table->string('rate_down_unit',4);
-            $table->integer('rate_up');
+            $table->float('rate_up');
             $table->string('rate_up_unit',4);
             $table->timestamps();
